Backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Machine Learning and Data Analysis
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.1.0
dtaidistance>=2.3.0

# Communication (Optional - only if using MQTT broker)
paho-mqtt>=2.0.0

# SMS Alerts (Optional - only if using Twilio)
twilio>=8.0.0

# Utilities
python-dotenv>=1.0.0

# Testing (Development only)
pytest>=7.4.0
pytest-asyncio>=0.21.0
